refactor: Step 1 in refactoring coursegraph for multiple backends
This work is incomplete, but makes progress towards being able to use ClickHouse or Neo4J as the coursegraph backend. It is untested in the current form and both settings and documentation updated.
